chore(knockdog): 디렉토리 구조 및 파일 네이밍 변경
