+++ apps/knockdog/src/entities/user/api/verifyEmail.ts
@@ -1,0 +1,13 @@
+import { api, ApiResponse } from '@shared/api';
+
+interface VerifyEmailParams {
+  code: string;
+  email: string;
+}
+
+/** `POST` - 이메일 인증 API */
+const postVerifyEmail = async (params: VerifyEmailParams): Promise<ApiResponse> => {
+  return await api.post('auth/email/verify', { json: params }).json<ApiResponse>();
+};
+
+export { postVerifyEmail };
