+++ apps/knockdog/src/entities/user/index.ts
@@ -1,5 +1,6 @@
-export type { User, UserAddress, UserStatus, UserAddressType } from './model';
+export { postVerifyEmail } from './api/verifyEmail';
 
-export { USER_STATUS, USER_ADDRESS_TYPE, USER_ADDRESS_TYPE_KR, useUserStore } from './model';
+export { USER_STATUS, USER_ADDRESS_TYPE, USER_ADDRESS_TYPE_KR } from './model/constants/user';
+export { useUserStore } from './model/store/userStore';
 
-export { postLogin, postLogout, postWithdraw } from './api';
+export type { User, UserAddress, UserStatus, UserAddressType } from './model/user';
